upgrade

+++ exercises/functions/functions2.rs
@@ -1,7 +1,5 @@
 // functions2.rs
-//
-// Execute `rustlings hint functions2` or use the `hint` watch subcommand for a
-// hint.
+// Execute `rustlings hint functions2` or use the `hint` watch subcommand for a hint.
 
 
 
@@ -9,7 +7,7 @@ fn main() {
     call_me(3);
 }
 
-fn call_me(num: usize) {
+fn call_me(num: i32) {
     for i in 0..num {
         println!("Ring! Call number {}", i + 1);
     }
